Support for building manopages with pod2man
Now there is a "man" target in the Makefile which allows for building the
manual pages for the included utilities. From now on every utility added
should have a manual page, at least a brief one, unless there is a good
reason to not do so. Support for installing manual pages is still lacking,
though.